finish homePage
