+++ src/pages/HomePage/components/BackgroundRectangle.js
@@ -2,7 +2,7 @@ import styled from '@emotion/styled'
 
 const BackgroundRectangle = styled.div`
   width: 70vw;
-  height: 200px;
+  height: 300px;
   background-color: ${({ theme }) => theme.secondary_850};
 `
