+++ src/pages/HomePage/components/Curve.js
