+++ src/pages/HomePage/components/Session.js
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled'
+
+const Session = styled.div`
+  min-height: 80vh;
+  position: relative;
+  padding-top: 2rem;
+`
+
+export default Session
